Revert "fix load configs"
This reverts commit f6834d0fbc731c725801b736e9f0788fec33cd70.

Revert "Fix;"

This reverts commit eaa0d9b5cad0e29e3ee0adba38f3f3c1b6ed0409.

Revert "return values with fields;"

This reverts commit bc09fbf7c479bcbf4f8a0b6968b4171a6518a428.

Revert "very bad fix;"

This reverts commit e1105c6ec26ab765379cc9a1f1b6910349d5402e.

+++ DTF2020/Assets/ScriptsAnother/Utils/utilFunction.cs
@@ -163,7 +163,15 @@ public class utilFunction
 
     static string loadFile(string nameConfig)
     {
+        if (!Directory.Exists(Application.streamingAssetsPath))
+        {
+            return "";
+        }
         var dir = Path.Combine(Application.streamingAssetsPath, nameConfig + ".json");
+        if (!File.Exists(dir))
+        {
+            return "";
+        }
         return File.ReadAllText(dir);
     }
 
@@ -189,6 +197,10 @@ public class utilFunction
 
     static protected void saveFile(string data, string nameFile)
     {
+        if (!Directory.Exists(Application.streamingAssetsPath))
+        {
+            return; 
+        }
         var dir = Path.Combine(Application.streamingAssetsPath, nameFile + ".json");
         File.WriteAllText(dir, data);
     }
